google-adk-multiagent-app/src/requirements.txt
streamlit
google-cloud-bigquery
pandas
plotly
python-dotenv
requests